Clears up confusing language on removing rows
Just spent SOOOO long trying to figure this out.  If I have `getRowNodeId: data => data.primaryKey` and I want to remove the row that has the primary key `abc`, I was trying to pass the transaction object `{ remove: ['abc'] }` because I interpreted the docs as saying to do so (and also because that's the simpler approach).  Instead I really needed to pass `{ remove: [{primaryKey: 'abc'}] }`.  My frustration was exacerbated by the fact that the ag-grid documentation examples for the `remove` portion of the transaction all don't use `getRowNodeId` which would have immediately helped resolve my problem much earlier.  I suggest making an example that does so.

Feel free to change my wording here as you like, just wanted to raise the point that it could be much clearer.  Perhaps adding the type of RowNode to this portion of the docs would make it clearer as well?