Fix IE11 height issue
We noticed that in some cases, IE11 will not heed the
flexbox rule that flexbox items do not shrink below
their minimum content size, and will use flex-basis
instead. With
  flex: 1;
flex-basis is set to 0%, which then leads to grids
with a height of zero. Setting explicitly
  flex: 1 1 auto;
as done previously for another item in
https://github.com/ag-grid/ag-grid/commit/6bd222946c0d26a7a42a40f38bf4dff94bb3b30d
works around this.