Trigger onCellValueChanged on Paste
Issue #3364 

As of version 21.2.1 ag-grid no longer triggers onCellValueChanged when pasting in data. This is due to setValue being called with suppressCellValueChangedEvent=true in updateCellValue method inside clipboardService class.

Suggested fix is to remove true parameter and set suppressCellValueChangedEvent to undefined. Or to remove return statement in ag-grid-enterprise ValueService.prototype.setValue
`
if (suppressCellValueChangedEvent) {
    return;
}
`