Use rowDataChangeDetectionStrategy / deltaRowDataMode for pinned rows
Currently if you set the `rowDataChangeDetectionStrategy` or `deltaRowDataMode` props they are only used for the `rowData`.

I believe this should also be applied to `pinnedTopRowData` and `pinnedBottomRowData`.

The issue I am facing is that I have a pinned top row, where the data in each cell is a complex object. As I cannot set the change detection strategy for this row, the default change detection strategy `DeepValueCheck` is applied, which is extremely slow (4-5 seconds to render). I would like to be able to override this to use an `IdentityCheck` instead.