Allow ranged deselection.
A customer intended to select large numbers of rows, and also intended to allow
users to deselect parts of these ranges.

The existing functionality only supported adding rows to a range selection.
This feature allows the user to `click`, then `shift+click` in any
direction on the grid, allowing for ranged selection and deselection of nodes.

Because `RowNode.setSelected` may be triggered programatically (ie when updating
rows without a user interaction), we prevent the range selection from happening.

When the user toggles the selection of a row via checkbox, with the `shift` key
held down, we allow the range selection to occur.