AG-7214 - Use resize observer to update popup position when first open (#5705)
* AG-7214 - Use resize observer to update popup position when first open

Fix bug where popup would show underneath the grid because the calculation happens before the popup contents have rendered.

* AG-7214 - Only use resize observer when mouse tracking

When mouse tracking, the resize observer update adds jitteriness, so it is disabled. Usually mouse tracking will be used on mouse move (eg, tooltips), so the recalculation of the position is done when the mouse moves anyway.