AG-8132 - Don't play the example again if it is already playing
Happens when two intersection observers both send a `start()` eg, when integrated charts is intersecting it triggers a `start()` and when row grouping is not intersecting, it checks if other grids are on the screen and `start()`s them.

This is fixed at the `scriptRunner` level, so that other it can't happen again when called in other places.