AG-8132 - Remove integrated charts initial wait
Move integrated charts initialisation to `onFirstDataRendered` so that the script can initialise after that has happened. Makes it faster for integrated charts example to start up.

NOTE: Can't do this for row grouping because the integrated charts intersection observer triggers a start of the grid before it row grouping's `onFirstDataRendered`, so row grouping won't start if refreshed right on it.