RTI-1938 - Replace site url with window location when opening plunkr/codesandbox
For working with the IP addresses or other hostnames on local development.

Not implemented in `SystemJs.tsx` file as that file is built statically, and `window.location` can only be used on the client side. Rather than make everything require client side code for system js, just do it on clicking on opening in plunkr/codesandbox.