RTI-1980 Support relative example app location path for example worker
In the source code generated for Plunker examples, the `appLocation` base path will be `""`. This was causing problems for the Application Created Charts example, which was appending a worker script path with a leading `/`, which was causing an attempt to load the worker script from the domain root, rather than relative to the current script.