Properly remove outdated styles from DOM on theme change
By querying from `document.head`, outdated injected styles are left behind in the DOM when using a `themeStyleContainer` that is outside of `document.head` (like `document.body`, for example). The impetus to use such selectors could come from a desire to avoid layer collisions, such as were run into in #12035 . This proposed change simply widens the search to the entire document, ensuring that the stylesheets are found. Alternatively, the query could be made from `environment.eStyleContainer.querySelectorAll(...)`, which defaults to `document.head`, but if for some reason that container identifier changed, the query would fail (weird hypothetical...but I do like just using `document` regardless).